backend/requirements.txt
fastapi[standard]
uvicorn
httpx
mysql-connector-python
python-dotenv
bcrypt
google-genai